test: remove separate test files
